Fix logo not being visible in auxiliary views when sidebar is collapsed
