+++ src/app/root.less
@@ -20,6 +20,11 @@
     // magic
     --screentabs-height: 38px;
 
+    // floating logo settings
+    --floating-logo-width-darwin: 110px;
+    --floating-logo-width: 40px;
+    --floating-logo-height: var(--screentabs-height);
+
     // global colors
     --app-accent-color: rgb(88, 193, 66);
     --app-error-color: rgb(204, 0, 0);
